feat: improve mesh generation and manipulation
